POCOR-2522	- Bugs fixes after upgrade WIP

+++ plugins/Institution/src/Model/Table/StaffTable.php
@@ -471,7 +471,9 @@ class StaffTable extends AppTable {
 			$activeStatusId = $this->Workflow->getStepsByModelCode($positionTable->registryAlias(), 'ACTIVE');
 			$positionConditions = [];
 			$positionConditions[$this->Positions->aliasField('institution_id')] = $institutionId;
-			$positionConditions[$this->Positions->aliasField('status_id').' IN '] = $activeStatusId;
+			if (!empty($activeStatusId)) {
+				$positionConditions[$this->Positions->aliasField('status_id').' IN '] = $activeStatusId;
+			}
 			if (!empty($excludeArray)) {
 				$positionConditions[$this->Positions->aliasField('id').' NOT IN '] = $excludeArray;
 			}
@@ -820,12 +822,13 @@ class StaffTable extends AppTable {
 			$InstitutionRubrics->delete($value);
 		}
 
-		$InstitutionSubjectStaff = TableRegistry::get('Institution.InstitutionSubjectStaff');
-
-		$InstitutionSubjectStaff->deleteAll([
-			$InstitutionSubjectStaff->aliasField('staff_id') => $staffId,
-			$InstitutionSubjectStaff->aliasField('institution_subject_id') . ' IN ' => $subjectIdsDuringStaffPeriod
-		]);
+		if (!empty($activeStatusId)) {
+			$InstitutionSubjectStaff = TableRegistry::get('Institution.InstitutionSubjectStaff');
+			$InstitutionSubjectStaff->deleteAll([
+				$InstitutionSubjectStaff->aliasField('staff_id') => $staffId,
+				$InstitutionSubjectStaff->aliasField('institution_subject_id') . ' IN ' => $subjectIdsDuringStaffPeriod
+			]);
+		}
 
 		// this logic here is to delete the roles from groups when the staff is deleted from the school
 		try {
